[Xamarin.Android.Build.Tasks] require JDK 11 for API-31+ (#6084)
API-31 builds fail when using JDK 1.8 and using AndroidX with:

	/Library/Frameworks/Mono.framework/External/xbuild/Xamarin/Android/Xamarin.Android.Common.targets(1728,3):
	error JAVAC0000: java.lang.AssertionError: annotationType() : unrecognized Attribute name MODULE (class com.sun.tools.javac.util.SharedNameTable$NameImpl)

There are two plausible solutions to better handle this:

 1. Only require JDK 11 when API-31 and AndroidX are being used, or
 2. Always required JDK 11 for API-31+

As we believe *most* apps will be using AndroidX, there is little
point to attempting so support (1); (2) it is!

We must require JDK 11 for API-31+.

There are two code paths here: one for "legacy" Xamarin.Android, and
one for .NET 6.

Legacy is straightforward, as we already had a pattern for this, we
just update the `<ValidateJavaVersion/>` task.

For .NET 6, I checked the `$(TargetPlatformVersion)` property for now
and require JDK 11 in that case.  When API-31 is stable, we can
probably remove this code and just set this value by default:

	<MinimumSupportedJavaVersion Condition=" '$(MinimumSupportedJavaVersion)' == '' ">11.0</MinimumSupportedJavaVersion>

For our unit tests -- most of which would be solution (1) (API-31
and *no* AndroidX use), but many of which started to fail because we
went with solution (2) -- we did one of two things:

 1. Use a different `$(TargetFrameworkVersion)` and/or remove
    `$(TargetFrameworkVersion)` entirely, or

 2. Use a new `BuildTest.AssertTargetFrameworkVersionSupported()`
    method which `Assert.Ignore()`s the test if it would not
    pass otherwise.