[build-tools] setup a way to install .NET MAUI
We need a way for testing to install builds of .NET MAUI alongside a
local build of the Android workload.

This creates a new target such as:

    msbuild Xamarin.Android.sln -t:InstallMaui -p:MauiVersion=6.0.100-rc.1.1351

You will need to lookup the version of the
`Microsoft.NET.Sdk.Maui.Manifest-6.0.100` package you want on [Azure
DevOps][0] or nuget.org.

[0]: https://dev.azure.com/azure-public/vside/_packaging?_a=package&feed=xamarin-impl%40Local&package=Microsoft.NET.Sdk.Maui.Manifest-6.0.100&protocolType=NuGet&version=6.0.100-rc.1.1351%2Bsha.3fbb791e7-azdo.5078933