[Xamarin.Android.Build.Tasks] support $(AndroidUseLatestPlatformSdk) in binding projects
After API 31 is stable, binding projects with
`TargetFrameworkVersion=11.0` and `AndroidUseLatestPlatformSdk=true`
fail to build with:

    Task ResolveNuGetPackageAssets
    ...
    Microsoft.NuGet.targets(198,5): Your project does not reference "MonoAndroid,Version=v12.0" framework. Add a reference to "MonoAndroid,Version=v12.0" in the "TargetFrameworks" property of your project file and then re-run NuGet restore.

What happened is that `$(TargetFrameworkVersion)` was changed to 12.0
during `Build`, but not during `Restore`. So NuGet restored with
`MonoAndroid,Version=v11.0` instead of `MonoAndroid,Version=v12.0`.

I found that we have this setup for other project types, but not
binding projects:

    <!--
      NOTE:
      This target runs during Restore, and there is no $(RestoreDependsOn) property.
      There appears to be no other way to do this other than use BeforeTargets.
    -->
    <Target Name="_SetLatestTargetFrameworkVersionForPackageReference"
        Condition=" '$(AndroidUseLatestPlatformSdk)' == 'True' "
        BeforeTargets="_GetRestoreTargetFrameworksOutput"
        DependsOnTargets="_SetLatestTargetFrameworkVersion">
    </Target>

If we move this target to `Xamarin.Android.Legacy.targets`, it will be
imported by binding projects.

We also get the added benefit of it not being in .NET 6.
`$(AndroidUseLatestPlatformSdk)` is not supported in .NET 6, so we
should not have this MSBuild target there.