Speed up p/invoke override lookups
Instead of using the `robin_map` hash map for the internal p/invokes,
incurring startup time penalty as the map has to be instantiated and
populated with `std::string` instances, use a plain array of structures
describing each p/invoke entry.  Each structure contains the p/invoke
function name, a pointer to that function as well as a hash calculated
on the build time using the xxHash algorithm.  The arrays are sorted on
the hash value so that during run time we can use a quick binary search
to locate the right function for the requested p/invoke.

There are two sets of p/invokes - one for internal Xamarin.Android
p/invokes (corresponding to the `xa-internal-api` and `java-interop`
"library" names) and another for the shared libraries currently shipped
with DotNet6.  The former are fully configured on build time, since we
can directly specify pointers to the functions implementing a p/invoke.
The latter, however, require more work on the run time since we cannot
set the pointer to a function during build.  Instead, on first request
for a given p/invoke, we load the relevant library and look up the
correct symbol, storing the pointer in the array afterwards.

Both the internal and dotnet library names are matched without using
string comparison.  Instead, a hash of the library name passed from Mono
is generated and compared against pre-generated hashes to determine
which table is to be used for lookups.

Both the tables and library name hashes are generated by the
`generate-pinvoke-tables` C++ program which also serves as a build-time
test of the xxHash implementation (it includes a handful of simple
compile-time tests).  Alas, currently this program can be built only on
the Linux CI bots, since neither Windows nor macOS machines there
support enough of C++20.  For this reason, the generated file is
committed to the repository instead of generating it each time
`src/monodroid` is built.  However, when running on CI Linux machines,
we do compile the utility, generate the tables and compare the generated
file to the one committed to repository.  If the files differ the build
will error out and let us examine the differences (both the new
generated file and the diff are part of the build status archive).  More
work needs to be done in order to compare the public interfaces found in
the dotnet shared libraries against the table stored in the generator.

The generated tables allow us to not use the "slow" `robin_map<>` hash
path during application startup (MAUI apps use the DotNet p/invokes
during that phase).  Additionally, the `robin_map<>` cache now uses the
xxHash hashes as well for faster lookups.

In addition, both the embedded assembly mmap code as well as the DotNet
p/invoke entry code now use the GCC/clang `__atomic` interfaces to
atomically set pointers, which avoids having to use any form of locks.

A new lock guard class, `StartupAwereLock`, is used to take a lock only
after the startup phase of Xamarin.Android is complete and it's possible
that multiple threads may be calling the runtime routines.

All of the above changes improve application startup time on a Pixel 3
XL device by:

  * Plain XA sample: ~17ms faster
  * Hello MAUI sample: ~100ms faster