[Mono.Android] Fix incorrect Bluetooth enumification (#6214)
Fixes: https://developercommunity2.visualstudio.com/t/xamarin-android-bluetooth-the-advertisin/841102
Fixes: https://github.com/xamarin/xamarin-android/issues/2595

Update enumification of various
`Android.Bluetooth.LE.AdvertisingSetParameters.Builder` and
`Android.Bluetooth.BluetoothDevice` methods to use the
`Android.Bluetooth.BluetoothPhy` enum instead of the
`Android.Bluetooth.LE.ScanSettingsPhy` enum, introduced in e33eb53407.

Backwards compatibility is achieved by creating overloads for the
existing methods in `Additions`, and marking them as `[Obsolete]`.

The API "breakage" is because the `[Register]` attribute has moved
from the incorrect methods to the correct methods.  As the signature
change is an enum, the marshalled type is still an `int`, so moving
the `[Register]` should not cause issues.

For example:

	// Current
	[Register]
	public void DoSomething (BadEnum value) { ... }

	// Fixed
	[Register]
	public void DoSomething (GoodEnum value) { ... }

	[Obsolete ("Use DoSomething(GoodEnum)")]
	public void DoSomething (BadEnum value) => DoSomething ((GoodEnum)value);

Note: this is only generally save to do when `DoSomething()` is a
*non-`virtual`* method (i.e. Java `final` method).  Fortunately, this
is the case with the methods involved here.