[tests] simpler test case for org.apache.http.legacy
Context: https://github.com/xamarin/xamarin-android/pull/6263#issuecomment-912671300

We have a "flaky" test:

    Xamarin.Android.Build.Tests.BuildTest.MissingOrgApacheHttpClient("dx")
    Xamarin.ProjectTools.FailedBuildException : Build failure:
    ...
    /Library/Frameworks/Mono.framework/Versions/6.12.0/lib/mono/msbuild/Current/bin/NuGet.targets(131,5):
    error : Could not find file '/Users/runner/.nuget/packages/xamarin.googleplayservices.tasks/117.0.0/bsd2z4yg.43k'.
    [/Users/runner/work/1/s/bin/TestRelease/temp/MissingOrgApacheHttpClientdx/UnnamedProject.csproj]

We have a `[Retry(5)]` on this, which doesn't seem to help.

To simplify what this is testing, we don't actually need to bring in
Google Play Services & Xamarin.Build.Download.

I could create an `ApacheHttpClient.cs` file that uses the deprecated
org.apache.http APIs. This test still fails if I remove this from the
`AndroidManifest.xml`:

    <uses-library android:name="org.apache.http.legacy" android:required="false" />

I think this is sufficient to test what we need, and should hopefully
pass more reliably.