Special Char Support
Binutils 2.36 introduced [code][0] to deal extra long path names on
Windows which makes compilation (in our case using `as`) and linking
to break if the source/object files reside in a directory which has
special characters (in our case they were `テスト`):

    Xamarin.Android.Common.targets(1887,3): error XA3006: Could not compile native assembly file: typemaps.armeabi-v7a.s [SmokeTestBuildWithSpecialCharactersFalse\テスト\テスト.csproj]
      [aarch64-linux-android-as.EXE stderr] Assembler messages: (TaskId:187)
      [aarch64-linux-android-as.EXE stderr] Fatal error: can't create typemaps.arm64-v8a.o: Invalid argument (TaskId:187)

Downgrade binutils version to 2.35.2 which doesn't contain the offending
code block and will work for us, with relative paths. It will **still**
break with full paths if the path contains special characters, but since
our use case involves only relative paths, this is fine.

There is also a bug in aapt2 [1] which stops it processing a directory which has
certain special characters on windows. This will only appear on projects which
have library resources, such as Xamarin.Forms. The problem only appears when
processing a directory into a flata archive. Oddly it works file when processing a
file directly. In order to work around this we now generate a resource only `res.zip`
file when we extract the library resources. This `res.zip` is then used during the
generation of the `flata` archive. Weirdly, this works where directory transversal does
not.

Additionally it seems to be slightly quicker

Before
`581 ms  Aapt2Compile                               1 calls`

After
`366 ms  Aapt2Compile                               1 calls`

So the time taken to generate the `res.zip` is offset by the quicker execution of
`aapt2 compile`.

Finally, `java` tooling such as `zipalign` and `apksigner` also failed with the
special characters. This was because we were using full paths to the files
we were signing/aligning. Switching over to use relative paths fixes this particular
problem.

Add a set of Unit tests to make sure we can handle the special characters.

[0]: https://sourceware.org/git/?p=binutils-gdb.git;a=blob;f=bfd/bfdio.c;h=463b3879c52ba6beac47190f8eb0810b0c330e65;hb=HEAD#l124
[1] https://issuetracker.google.com/issues/188679588