[Xamarin.Android.Build.Tasks] introduce @(AndroidResourceId)
We currently have an issue where a class library:

1. Includes AndroidX & Google Material
2. Includes at least one `@(AndroidResource)` and uses the ID from C#.
3. `Resource.designer.cs` has 2,700+ fields.

2,700 fields is not a huge problem for a single class library. But
then the problem compounds:

1. Library A has 2,700 fields. It is a dependency of many other
   libraries.
2. Library B, C, and D have a dependency on A, and have at least one
   `@(AndroidResource)` that is used from C#.
3. Application A uses Library A, B, C, and D. Application A sets 2,700
   X 4 fields at startup -- over 10,000 fields.

We saw a case where a .NET MAUI app with SkiaSharp and Telerik set
90,000 fields at startup.

The solution here isn't easy. We can't just set
`$(AndroidGenerateResourceDesigner)` to `false` because the values are
used in C#. What we need is a way for "advanced users" to list which
fields their class library uses.

I propose:

    <AndroidResource Include="Resource.Drawable.foo" />
    <AndroidResource Include="Resource.Id.bar" />

If `@(AndroidResourceId)` is blank, you get the existing behavior.

In these names you should be able to omit `Resource.` if you like,
just to make things less error-prone.

After this change goes in, I would update SkiaSharp and .NET MAUI to
opt into this behavior.