[One .NET] fix $(ApplicationId) and //manifest@package at the same time
Fixes: https://github.com/xamarin/xamarin-android/issues/6302

If you try to use both `$(ApplicationId)` and `@package` in
`AndroidManifest.xml` at the same time, you hit errors during
deployment:

    Xamarin.Android.Common.Debugging.targets(611,5): error XA0132: The package was not installed. Please check you do not have it installed under any other user.
    If the package does show up on the device, try manually uninstalling it then try again.
    You should be able to uninstall the app via the Settings app on the device.

From the spec, the `AndroidManifest.xml` is *supposed* to be preferred.

I found the `<GetAndroidPackageName/>` MSBuild task is accidentally
preferring the `$(ApplicationId)` MSBuild property. I reordered the
logic and only use the property when no value is found in the
`AndroidManifest.xml`.

I could reproduce the issue in a test, which now passes.