[Xamarin.Android.Build.Tests] Fix GetMaxInstalledPlatform threading.
When we have unit tests running in parallel we sometimes get issues
where the method `GetMaxInstalledPlatform` returns a lower API level
than the max installed. This is because multiple tests are calling
the method at the same time. As a result they overwrite each other
as they are both scanning the directories. The results in some tests
getting a lower API level than expected becuase the `maxInstalled`
field has a value.