[WIP] Don't add the `Abi` metadata to satellite assemblies
Doing so would put them in the arch-specific blob, which is incorrect in
and of itself. Also, oddly enough, the assemblies would end up with
`Abi` only for the first architecture targetted by the app, which would
cause the arch-specific blob generation to fail because of different
number of assemblies in different architectures.