[build] Replace remaining 'netcoreapp3.1' usage with 'net6.0' (#6369)
Our build system requires .NET 6 to build many artifacts.  However,
there are some artifacts still building and running on .NET Core 3.1,
requiring that .NET Core 3.1 be installed as well.

Update `build-tools/xaprepare` and `tools/tmt` to use `net6.0`, so
that .NET Core 3.1 is no longer required to build xamarin-android.

One complication is that the default version of Mono available on the
Azure Pipelines Mac image is not new enough to build `net6.0` projects.
As part of our build, we download and install the latest stable Mono
to fix this.  However, this step is performed by `xaprepare`, which
previously could be compiled by the default Mono (`netcoreapp3.1`)
but now cannot (`net6.0`).

Fix this by building `xaprepare` with .NET 6, not Mono.
Introduce a new `run-xaprepare.yaml` template to assist this.

Building `xaprepare` with .NET 6 also requires updating the
[`//UsingTask/Code/@Source`][0] attribute value, used with
Inline Tasks, to use Unix-style paths instead of Windows-style paths.
This is necessary because the [`@Source` attribute is used as-is][1]
with `File.ReadAllText()`; no normalization is performed.  Unix-style
paths work on Windows, while Windows-style paths won't work on Unix,
so Unix-style paths must be used everywhere in order for this to work.

[0]: https://docs.microsoft.com/en-us/visualstudio/msbuild/msbuild-inline-tasks?view=vs-2019#code-element
[1]: https://github.com/dotnet/msbuild/blob/bbb9655b007be6d079985f3a7ec14f5d82a18f64/src/Tasks/RoslynCodeTaskFactory/RoslynCodeTaskFactory.cs#L404-L430