[One .NET] prefer /manifest/@android:versionCode
Fixes: https://github.com/dotnet/sdk/issues/21614

In .NET 6, you can't currently put `versionCode` in
`AndroidManifest.xml`, as it is always overwritten by a default:

    <ApplicationVersion Condition=" '$(ApplicationVersion)' == '' ">1</ApplicationVersion>

We should actually check if the value is present in
`AndroidManifest.xml` and only set if blank.

The other new MSBuild properties don't appear to suffer from this
problem:

* `$(ApplicationId)`
* `$(ApplicationTitle)`
* `$(ApplicationDisplayVersion)`

This is because we don't default these properties to a value when they
are blank.

I added a new test for this scenario.