[xaprepare] Replace `msbuild` with `dotnet build`
As part of a broader effort to `dotnet build` all the things, all usage
of `msbuild` in xaprepare has been replaced with `dotnet build`.

I've also attempted to simplify local and external prepare steps, as
there seemed to be a lot of duplicated restore logic in the various
`make` and `msbuild` invocations in those steps.