[ci] Fix DetermineApplicableTests build step
The test stages that run in a given PR are selected dynamically after
checking which files have changed in the PR.  This behavior relies on
the [output variables][0] Azure Pipelines feature.  In order to access
the output variable in other stages, the step that sets it must use the
`name:` property.

Fix output variable setting from the xaprepare yaml template by adding
an option `name` parameter.

[0]: https://docs.microsoft.com/en-us/azure/devops/pipelines/process/variables?view=azure-devops&tabs=yaml%2Cbatch#use-output-variables-from-tasks